feat: modules- add modale
