+++ src/blocks/profile/__avatar-container/profile__avatar-container.css
@@ -16,3 +16,4 @@
   .profile__avatar-container {
     margin: 0 auto 26px;
   }
+}
