+++ src/components/index2.js
@@ -1,55 +1,49 @@
-const profileOpenPopupButton = document.querySelector(".profile__edit-button");
-const profilePopup = document.querySelector(".popup-edit");
-const profileCloseButton = document.querySelector(".popup-edit__close");
-const popupCard = document.querySelector(".popup-card");
-const buttonAddCard = document.querySelector(".profile__add-button");
-const buttonCloseCard = document.querySelector(".popup-card__close");
-const cards = document.querySelector(".card");
+import '../pages/index.css';
 const popupAddedCardButton = document.querySelector(".popup-card__button");
-const popupPhoto = document.querySelector(".popup-photo");
-const popupPhotoClose = document.querySelector(".popup-photo__close");
 const photoLink = document.querySelector(".card__image");
 const photoName = document.querySelector(".card__name");
-const popupPhotoName = document.querySelector(".popup-photo__name");
-const popupPhotoImgLink = document.querySelector(".popup-photo__img");
-
-const nameInput = document.querySelector(".form__text_input_name");
-const jobInput = document.querySelector(".form__text_input_job");
 const profileInfo = document.querySelector(".profile__info");
 
 const profileName = document.querySelector(".profile__name");
-const profileText = document.querySelector(".profile__text");
-const profileForm = document.querySelector(".popup_form-profile");
 const popupFormCard = document.querySelector(".popup__form-card");
 
-// const editProfile = document.querySelector('.profile__avatar-container');
-// const popupProfileAvatar = document.querySelector('.popup-profile');
-// const popupProfileClose = document.querySelector('.popup-profile__close')
+import {
+  profileOpenPopupButton,
+  profilePopup,
+  profileCloseButton,
+  popupCard,
+  buttonAddCard,
+  buttonCloseCard,
+  popupPhoto,
+  popupPhotoClose,
+  nameInput,
+  jobInput,
+  profileName,
+  profileText,
+  profileForm,
+  openPopup,
+  closePopup,
+  handleProfileFormSubmit
+} from './modal';
 
-function openPopup(popup) {
-  popup.classList.add("popup_opened");
+import {
+  popupCard,
+  cards,
+  popupPhoto,
+  popupPhotoName,
+  popupPhotoImgLink,
+  templateCard,
+  nameCard,
+  linkCard,
+  render,
+  createCard,
+  renderinitialCards,
+  addCard,
+  setDeleteHandler,
+  handleCardDelete
+} from './card'
 
-  function closeCurTargetPopup(evt) {
-    if (evt.target === evt.currentTarget) {
-      closePopup(popup);
-      popup.removeEventListener("click", closeCurTargetPopup);
-    }
-  }
-
-  function closeKeyPopup(evt) {
-    if (evt.key === "Escape") {
-      evt.preventDefault();
-      closePopup(popup);
-      document.removeEventListener("keydown", closeKeyPopup);
-    }
-  }
-  popup.addEventListener("click", closeCurTargetPopup);
-  document.addEventListener("keydown", closeKeyPopup);
-}
-
-function closePopup(popup) {
-  popup.classList.remove("popup_opened");
-}
+// ----------------------
 
 profileOpenPopupButton.addEventListener("click", function () {
   openPopup(profilePopup);
@@ -70,23 +64,6 @@ buttonCloseCard.addEventListener("click", function () {
 popupPhotoClose.addEventListener("click", function () {
   closePopup(popupPhoto);
 });
-
-// editProfile.addEventListener('click', function () {
-//   openPopup(popupProfileAvatar);
-// });
-// popupProfileClose.addEventListener('click', function () {
-//   closePopup(popupProfileAvatar)
-// });
-
-// Открытие и закрытие попап
-
-function handleProfileFormSubmit(evt) {
-  evt.preventDefault();
-  profileName.textContent = nameInput.value;
-  profileText.textContent = jobInput.value;
-  closePopup(profilePopup);
-}
-
 profileForm.addEventListener("submit", handleProfileFormSubmit);
 
 // добавление информации в профиль
@@ -186,7 +163,68 @@ function handleCardDelete(event) {
 
 render();
 
-import {
-  enableValidation
-} from '.validate.js';
+// --------------------------------------------валидация-----------------------------------------
+
+const showInputError = (formElement, inputElement, errorMessage) => {
+  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+  inputElement.classList.add("form__input_type_error");
+  errorElement.textContent = errorMessage;
+  errorElement.classList.add("form__input-error_active");
+};
+
+const hideInputError = (formElement, inputElement) => {
+  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+  inputElement.classList.remove("form__input_type_error");
+  errorElement.classList.remove("form__input-error_active");
+  errorElement.textContent = "";
+};
+
+const checkInputValidity = (formElement, inputElement) => {
+  if (!inputElement.validity.valid) {
+    showInputError(formElement, inputElement, inputElement.validationMessage);
+  } else {
+    hideInputError(formElement, inputElement);
+  }
+};
+
+const hasInvalidInput = (inputList) => {
+  return inputList.some((inputElement) => {
+    return !inputElement.validity.valid;
+  });
+};
+
+const toggleButtonState = (inputList, buttonElement) => {
+  if (hasInvalidInput(inputList)) {
+    buttonElement.classList.add("button_inactive");
+    buttonElement.disabled = true
+  } else {
+    buttonElement.classList.remove("button_inactive");
+    buttonElement.disabled = false
+  }
+};
+
+const setEventListeners = (formElement) => {
+  const inputList = Array.from(formElement.querySelectorAll(".form__input"));
+  const buttonElement = formElement.querySelector(".form__submit")
+
+  toggleButtonState(inputList, buttonElement);
+  inputList.forEach((inputElement) => {
+    inputElement.addEventListener("input", function () {
+      checkInputValidity(formElement, inputElement);
+      toggleButtonState(inputList, buttonElement);
+    });
+  });
+};
+
+const enableValidation = () => {
+  const formList = Array.from(document.querySelectorAll(".form"));
+  formList.forEach((formElement) => {
+    formElement.addEventListener("submit", (evt) => {
+      evt.preventDefault();
+    });
+    setEventListeners(formElement);
+  });
+};
 enableValidation();
+
+//   -------------------------------------------------------------------------------------------
