+++ src/components/modal.js
@@ -6,10 +6,8 @@ const buttonAddCard = document.querySelector(".profile__add-button");
 const buttonCloseCard = document.querySelector(".popup-card__close");
 const popupPhoto = document.querySelector(".popup-photo");
 const popupPhotoClose = document.querySelector(".popup-photo__close");
-
 const nameInput = document.querySelector(".form__text_input_name");
 const jobInput = document.querySelector(".form__text_input_job");
-
 const profileName = document.querySelector(".profile__name");
 const profileText = document.querySelector(".profile__text");
 const profileForm = document.querySelector(".popup_form-profile");
@@ -39,8 +37,6 @@ function closePopup(popup) {
   popup.classList.remove("popup_opened");
 }
 
-// Открытие и закрытие попап
-
 function handleProfileFormSubmit(evt) {
   evt.preventDefault();
   profileName.textContent = nameInput.value;
@@ -63,8 +59,6 @@ export {
   profileText,
   profileForm,
   openPopup,
-  closeCurTargetPopup,
-  closeKeyPopup,
   closePopup,
   handleProfileFormSubmit
 }
