+++ webpack.config.js
@@ -17,9 +17,9 @@ module.exports = {
   mode: 'development',
   devServer: {
     contentBase: path.resolve(__dirname, './dist'),
-    open: true,
     compress: true,
-    port: 8080
+    port: 8080,
+    open: true,
   },
   module: {
     rules: [{
